# api/openapi.yaml
openapi: 3.0.3
info:
  title: ОбъектМаркет API
  version: 1.0.0
  description: 'API учебной системы закупки строительных материалов. Суммы — в копейках; реальные платежи и GPS отсутствуют. Демонстрационная доставка доступна только для строк адреса вида «Астрахань, улица, дом» или «г. Астрахань, улица, дом».'
servers:
  - url: /api/v1
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: om_session
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code: {type: string}
        message: {type: string}
...
    Calculation:
      type: object
      properties:
        formula: {type: string}
        calculatedConsumption: {type: number}
        consumptionUnit: {type: string}
        packages: {type: integer}
        packSize: {type: number}
  responses:
    Error:
      description: Ошибка с кодом и сообщением
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
paths:
...
  /projects:
    get:
      summary: Объекты покупателя
      security: [{session: []}]
      responses: {'200': {description: Список объектов}}
    post:
      summary: Создать объект
      security: [{session: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, address]
              properties:
                name: {type: string}
                address: {type: string}
                stages: {type: array, items: {type: object}}
      responses: {'201': {description: Новый объект}}
  /projects/{id}:
    parameters: [{name: id, in: path, required: true, schema: {type: string, format: uuid}}]
    get:
      summary: Объект и список потребностей
      security: [{session: []}]
      responses: {'200': {description: Объект}, '404': {$ref: '#/components/responses/Error'}}
    patch:
      summary: Изменить свой объект
      security: [{session: []}]
      responses: {'200': {description: Изменённый объект}}
  /projects/{id}/items:
    post:
      summary: Добавить позицию потребности
      security: [{session: []}]
      parameters: [{name: id, in: path, required: true, schema: {type: string, format: uuid}}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [productId, quantity]
              properties:
                productId: {type: string, format: uuid}
                quantity: {type: integer, minimum: 1}
                stageDate: {type: string, format: date}
      responses: {'201': {description: Позиция добавлена}}
  /calculators/tiles:
    post:
      summary: Расчёт плитки с округлением до упаковки
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [areaM2, tileAreaM2, tilesPerPack]
              properties:
                areaM2: {type: number, minimum: 0, exclusiveMinimum: true}
                tileAreaM2: {type: number, minimum: 0, exclusiveMinimum: true}
                tilesPerPack: {type: integer, minimum: 1}
                wastePercent: {type: number, minimum: 0, default: 10}
      responses: {'201': {description: Формула, расход и число упаковок, content: {application/json: {schema: {$ref: '#/components/schemas/Calculation'}}}}}
  /calculators/paint:
    post:
      summary: Расчёт краски с учётом слоёв
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [areaM2, rateLPerM2, packSizeL]
              properties:
                areaM2: {type: number, minimum: 0, exclusiveMinimum: true}
                rateLPerM2: {type: number, minimum: 0, exclusiveMinimum: true}
                coats: {type: integer, minimum: 1, default: 2}
                packSizeL: {type: number, minimum: 0, exclusiveMinimum: true}
                wastePercent: {type: number, minimum: 0, default: 10}
      responses: {'201': {description: Формула, расход и число упаковок, content: {application/json: {schema: {$ref: '#/components/schemas/Calculation'}}}}}
  /calculators/dry-mix:
    post:
      summary: Расчёт сухой смеси по толщине слоя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [areaM2, layerMm, rateKgPerM2Mm, packSizeKg]
              properties:
                areaM2: {type: number, minimum: 0, exclusiveMinimum: true}
                layerMm: {type: number, minimum: 0, exclusiveMinimum: true}
                rateKgPerM2Mm: {type: number, minimum: 0, exclusiveMinimum: true}
                packSizeKg: {type: number, minimum: 0, exclusiveMinimum: true}
                wastePercent: {type: number, minimum: 0, default: 10}
      responses: {'201': {description: Формула, расход и число упаковок, content: {application/json: {schema: {$ref: '#/components/schemas/Calculation'}}}}}
...
  /geo/route:
    get:
      summary: Маршрут 2ГИС или схематический маршрут демо
      parameters:
        - {name: fromLon, in: query, required: true, schema: {type: number}}
        - {name: fromLat, in: query, required: true, schema: {type: number}}
        - {name: toLon, in: query, required: true, schema: {type: number}}
        - {name: toLat, in: query, required: true, schema: {type: number}}
      responses: {'200': {description: 'Источник, координаты [долгота, широта], дистанция и время'}, '502': {$ref: '#/components/responses/Error'}}